Lagt til antall i faktura_linje tabellen
* Lagt til antall som lagres i database

* Lagt til antall som sendes videre til kafka køen

+++ src/main/kotlin/no/nav/faktureringskomponenten/service/integration/kafka/dto/FakturaBestiltLinjeDto.kt
@@ -4,7 +4,7 @@ import java.math.BigDecimal
 
 data class FakturaBestiltLinjeDto(
     val beskrivelse: String,
-    val antall: Double,
+    val antall: BigDecimal,
     val enhetspris: BigDecimal,
     val belop: BigDecimal,
 )
